+++ integration/spark/spark32/src/main/java/io/openlineage/spark32/agent/lifecycle/plan/AlterTableCommandDatasetBuilder.java
@@ -6,6 +6,7 @@
 package io.openlineage.spark32.agent.lifecycle.plan;
 
 import io.openlineage.client.OpenLineage;
+import io.openlineage.client.dataset.DatasetCompositeFacetsBuilder;
 import io.openlineage.client.utils.DatasetIdentifier;
 import io.openlineage.spark.agent.util.PlanUtils;
 import io.openlineage.spark.api.AbstractQueryPlanOutputDatasetBuilder;
@@ -84,20 +85,21 @@ public class AlterTableCommandDatasetBuilder
     }
 
     OpenLineage openLineage = context.getOpenLineage();
-    OpenLineage.DatasetFacetsBuilder builder =
-        openLineage
-            .newDatasetFacetsBuilder()
-            .schema(PlanUtils.schemaFacet(openLineage, schema))
-            .lifecycleStateChange(
-                openLineage.newLifecycleStateChangeDatasetFacet(lifecycleStateChange, null))
-            .dataSource(PlanUtils.datasourceFacet(openLineage, di.get().getNamespace()));
+    DatasetCompositeFacetsBuilder builder = new DatasetCompositeFacetsBuilder(openLineage);
+    builder
+        .getFacets()
+        .schema(PlanUtils.schemaFacet(openLineage, schema))
+        .lifecycleStateChange(
+            openLineage.newLifecycleStateChangeDatasetFacet(lifecycleStateChange, null))
+        .dataSource(PlanUtils.datasourceFacet(openLineage, di.get().getNamespace()));
 
     if (includeDatasetVersion(event)) {
       Optional<String> datasetVersion =
           CatalogUtils3.getDatasetVersion(
               context, resolvedTable.catalog(), resolvedTable.identifier(), table.properties());
       datasetVersion.ifPresent(
-          version -> builder.version(openLineage.newDatasetVersionDatasetFacet(version)));
+          version ->
+              builder.getFacets().version(openLineage.newDatasetVersionDatasetFacet(version)));
     }
     return Collections.singletonList(outputDataset().getDataset(di.get(), builder));
   }