+++ integration/spark/spark33/src/main/java/io/openlineage/spark33/agent/lifecycle/plan/ReplaceIcebergDataDatasetBuilder.java
@@ -6,7 +6,10 @@
 package io.openlineage.spark33.agent.lifecycle.plan;
 
 import io.openlineage.client.OpenLineage;
+import io.openlineage.client.OpenLineage.OutputDataset;
+import io.openlineage.client.dataset.DatasetCompositeFacetsBuilder;
 import io.openlineage.spark.api.AbstractQueryPlanOutputDatasetBuilder;
+import io.openlineage.spark.api.DatasetFactory;
 import io.openlineage.spark.api.OpenLineageContext;
 import io.openlineage.spark3.agent.utils.DataSourceV2RelationDatasetExtractor;
 import io.openlineage.spark3.agent.utils.DatasetVersionDatasetFacetUtils;
@@ -52,22 +55,27 @@ public class ReplaceIcebergDataDatasetBuilder
     if (!(replace.table() instanceof DataSourceV2Relation)) {
       return Collections.emptyList();
     }
+    DatasetFactory<OutputDataset> datasetFactory = outputDataset();
     DataSourceV2Relation table = (DataSourceV2Relation) replace.table();
 
-    final OpenLineage.DatasetFacetsBuilder datasetFacetsBuilder =
-        context.getOpenLineage().newDatasetFacetsBuilder();
-    datasetFacetsBuilder.lifecycleStateChange(
-        context
-            .getOpenLineage()
-            .newLifecycleStateChangeDatasetFacet(
-                OpenLineage.LifecycleStateChangeDatasetFacet.LifecycleStateChange.OVERWRITE, null));
+    final DatasetCompositeFacetsBuilder datasetFacetsBuilder =
+        datasetFactory.createCompositeFacetBuilder();
+    datasetFacetsBuilder
+        .getFacets()
+        .lifecycleStateChange(
+            context
+                .getOpenLineage()
+                .newLifecycleStateChangeDatasetFacet(
+                    OpenLineage.LifecycleStateChangeDatasetFacet.LifecycleStateChange.OVERWRITE,
+                    null));
 
     if (includeDatasetVersion(event)) {
-      DatasetVersionDatasetFacetUtils.includeDatasetVersion(context, datasetFacetsBuilder, table);
+      DatasetVersionDatasetFacetUtils.includeDatasetVersion(
+          context, datasetFacetsBuilder.getFacets(), table);
     }
 
     return DataSourceV2RelationDatasetExtractor.extract(
-        outputDataset(), context, table, datasetFacetsBuilder);
+        datasetFactory, context, table, datasetFacetsBuilder);
   }
 
   @Override